NP-2524 Update Organization Photo added
